Refine add_contact
When the contact already exists in the bucket, add_contact reorders the contact.

This function adds the contact even if the bucket is full, so the number of
items in a bucket can be more than bucket size. Instead of limiting the maximum
bucket size, add_contact returns a contact to be validated if the bucket is
full.
The caller of add_contact method must validate the contact.