Rename Address to SocketAddr
Currently, there are two kinds of Address. One is network socket
address, which this patch renamed. The other is a public address
generated by key-pair.

It is distinguished because only network package uses socket address and
network package only uses socket address. But the next patch starts to
use a public address in network package, and this duplication makes code
hard to read.