Simplify Session through removing unready session
The session is product of a secret and nonce.
The unready session was introduced since only secret exists while
handshaking. The previous model was creating session with secret and
making it ready with nonce. But it just makes code complicate.
This patch makes session be always ready. In other words, there must be
secret and nonce to make session.