Distinguish the error about payment and fee
Currently, the same error is returned when the balance is lower than fee
or payment.
But these two errors must be distinguished since the parcel that has an
insufficient fee is an invalid parcel, and it cannot be added on the
block while the transaction has not enough balance for payment is a
valid transaction.