Make NodeId convertible to a SocketAddr
Currently, it's not possible to get SocketAddr from NodeId since NodeId
is hash of a socket address. It gives a burden to keep the map from
NodeId to SocketAddr.

This patch makes NodeId covertible to a SocketAddr.

The conversion is implemented by IntoSocketAddr trait instead of Into
trait to distinguish it from other conversions.