Simplify Kademlia extension
CodeChain's Kademlia implementation doesn't need the original kademlia's
routing table.

The original routing table is optimized to find the closest nodes from
the local. But CodeChain's implementation uses it to find the closest
nodes from some remote nodes.

And the bucket refresh is not used since CodeChain uses TCP and rely on
it to heart-beat remote nodes.