Make state two-level trie.
This change prepares the sharding. This patch makes state stored in a
two-level trie.
The first trie is the top-level state which has account and shard.
The second trie is the shard-level state which has asset scheme and
asset.

This patch doesn't introduce sharding. It just changes the state. It can
be seen that there is only one shard. But it needs more changes to
implement multiple shards.