Change the way to create the address of top level item
Currently, CodeChain calculates 31 bytes blake2b hash and adds a prefix
to get the address of the top level item.

This patch changes the way how CodeChain calculates the address in the
same way as CodeChain calculates the address of shard level items. More
specifically, this patch makes CodeChain use the 32 bytes hash whose
first byte is replaced by the prefix as the address.