Remove World
Currently, there is a world in the shard and world determines who can
mint new assets.
It removes world and shard user determines who can mint new assets.
So minting is possible if the address is the user of the Shard.
And there is a free shard that has no users. Anyone can mint in the free
shard.

And this patch also removes ShardMetadata, which stores the number of
shards, because it has no information.