Make all TopLevelState have a dedicated StateDB
I had have a plan to make all TopLevelState share a global cache with
StateDB. But it's difficult because StateDB itself cannot distinguish
whether the cache sees the best bloc or not.
So I changed a plan to make all state have different StateDB and merging
the cache of the best state to the cache of Client.