Reset Tendermint's proposal field when the round is increased
Since the proposal field means current round's proposal hash, a node
does not create a new proposal if the value is set. However, the
proposal value was not reset when the round is changed. So there was a
bug which a node could not create a proposal if the node received a
proposal in the previous rounds.

This fixes #983