Merge Parcel and Transaction in rpc module
This patch makes the term Parcel obsolete.

And it introduces the concept of transaction id to represent the hash
without approvals and signature.
After this patch, the transaction hash is the hash of a signed
transaction, and the transaction id is a hash of the transaction
payload. It's used by AssetTransfer to point the asset. So the only
transactions about asset have the id.