Remove self.move_to_step in the Tendermint's restore and backup with the new step
We saved the Tendermint's state before changing the state and restored
the previous state and votes with hope that the restoring votes will
successfully recreate all the state.

But this causes some issues.

1. Double vote / proposal problem

While we call the `move_to_step` in the previous step, it could create
a new vote/proposal although the node created the other vote/propose in
the same height and view.

And, while restoring the votes, `move_to_step` also could be called
before all votes are restored. This also causes double vote/proposal.

2. Height / Step mismatch

When we move to the next height, we first update the height and update
the step later. We were saving the changed height and unchanged step.

Fix the issues by not calling the `move_to_step` function in the restore
and saving the changed step.

In this way, we could restore the valid prevote lock and does not
creates a double vote/proposal.

This fixes the below issues.
https://github.com/CodeChain-io/codechain/issues/1234
https://github.com/CodeChain-io/codechain/issues/1249