Add a type wrapper for a password
Currently, crypto module uses &str for a password. It makes a debugger
print the password on a back trace. The wrapper cannot protect from
memory reading since it's still on a continuous memory. But it will help
to prevent the users leak their password unintentionally when they
report a bug.
Of course, using a prettier in debugger cannot be prevented. This is
just a minimal action.