Forbid non-active user be registered as special users
The active account means the account that has ever had the balance. In
other words, it's the account that has non-zero balance or seq. This
term is introduced in this commit but the concept is already used. The
public key of the account that has non-zero balance or seq cannot be
used as a regular key.

This commit forbids the non-active account becoming the shard users,
shard owners, the administrator of the asset or the approver. If it's
allowed the regular account to be special users, criteria for which of
the master account and the regular account becomes a special user is not
intuitive. It's a restriction to make the logic simple.