Replace EpochTransition with TermTransition
`is_epoch_begin` flag which is determined by the result of
`epoch_tranistion` call doesn't make any differences whether it is true
or false since `on_new_block`, `on_epoch_begin` does nothing on it. So
it is safe to remove and to replace with `is_term_begin` which also does
nothing for now for the further implementation.
Inserted `EpochTransition` is not read anymore and inserting it does not
change state root. So removing it doesn't introduce breaking change unless
reusing the key index.