Pay block proposal rewards in a term at once
Currently, the block reward is paid per block; this patch makes the
reward is paid at the end of the term if the term seconds is determined.

Solo and Tendermint are different. Solo pays the fee at the end of the
term, but Tendermint pays the fee at the end of the next term.