Use the genesis network id at the email title
The best way is using the current network id, but CodeChain cannot know
the current network id when creating an email alarm.
Instead of the current network id, it can get the genesis network id
easily. Because there is no use case to change the network id yet, it
would be a fine solution for now.