Fix the bug that checks votes inappropriately
Current code uses the weights of validators to count vote, but it should
be the delegations it had received. The weight should be used for
selecting the proposer only.

So, this patch makes the states store the amount of the delegations of
validators.

And it also fixes one more bug, the current code doesn't reset the
weights. So this patch makes it reset the weights of the validators when
there are no validators that has weights.