Do not request older headers if re-organization is not occurred
Header downloader has "best_hash," which is peer's best hash, and
"pivot," which is the best hash of received headers from a peer.

When header downloader receives headers from a peer, it checks the
first header's hash and the pivot's hash. If they are not the same,
the downloader concludes that there is a reorganization and request
older headers.

However, if the pivot is changed between header request and response,
the header downloader requests older hashes even though there is no
re-organization. This commit changes the header downloader check
re-organization more concisely.