Update the best block when best proposal header is updated
In Tendermint best block needs precommits in the next
header. Currently, the best block is updated when the best proposal
header's previous block is imported. However, the best block is not
updated when a proposal block's next header is imported.

This commit updates best block when the next header is imported.