Call new_blocks even though block verification queue is not empty
new_blocks was called when the block verification queue is empty. The
purpose of the behavior is not clear. Maybe it was used to
backpressure for block sync extension. Since we have another block
sync flow control and tendermint assumes that new_blocks is always
called, this commit changed the behavior of new_blocks to be always called.