Do not generate a seal if the block is generated from a past view
If generating a block takes too much time, the view could be changed
before the miner module requests the signature. If the Tendermint
module receives a signature request of an old view, it should ignore
the message. Before this commit, the Tendermint module was crashing
when it gets a signature request from an old view.