Fix the condition to check whether the new block could change the best block
When a new block is imported in Tendermint consensus, CodeChain checks
the conditions below to judge whether it is safe to change the best
block:

First, whether the score is higher than before. Second, whether the
height of a new block is higher than the finalized block's
height.

This commit fixes the code that is calculating the finalized
block's height erroneously.