Enable all grandchild of the best block to become a best proposal block
Before this commit, only the child of the best block and the child of
the best proposal block could be a new best proposal block. A block
that is a grandchild of the best block but not the child of the best
proposal block should be able to become a best proposal block.

After this commit, CodeChain checks whether a new block is the best
block's child or grandchild.