Add get_version and set_version
When the data structure of data that is saved in the KVDB is changed,
CodeChain should migrate the data. To identify the stored data
structure, we need to keep a version number. This commit adds code to
save and load a version number. To make the DB easier to manage, this
code saves versions in the same column with the same prefix.