Rename total_score to nonce in sync message
When CodeChain is synchronized by snapshot sync, `total_score`
doesn't mean the accumulated score from the genesis block.
To handle this, the sync extension is updated to not rely on
`total_score` for deciding whether a peer is leading or not, and
use the value only as a monotonically increasing nonce.