Do not assert the finalized_view while restoring
In the Tendermint restoring process, if the backup state is the Commit
step, CodeChain set the step to the Precommit step and handles the
votes. The purpose of the behavior is that calling functions that are
called when enters the Commit state.

Changing the "step" in the restoring process is fragile. It is easy to
miss to set some variables. The `finalized_view_of_current_block`
variable should be changed when the step is changed. It was lost
before.

This commit fixes the problem by ignoring the assertion check in the
restore process. Although it is not a perfect solution, it is
consistent. There is some code already do differently in the restore
process.