Reduce the failure rate of testing total shutdown
Currently, the test frequently failed.
This test changes the validator set to the nodes which are not started
yet and starts all new validators when the term changed.

What I found is they are far behind than the best chain, and they cannot
create a block until timeout because they are synchronizing.

I think it's because of the slow synchronization, but fast
synchronization is hard to achieve in the near future. So this patch
made the new validator nodes keep alive while creating the blocks that
set initial states.

It may not be a perfect solution, but it reduces the test failure rates
dramatically.