Remove Trie::get_with
We prefer using an interface instead of a concrete object to make code
open and generic.
But the current code reveals TrieDB that is an implementation of Trie
because the Trait which has a method having generic parameters cannot be
an object.

This commit removes the get_with method and replaces TrieDB with Trie.