Make CodeChain makes blocks slowly if RUN_ON_TEST is set
This will make CodeChain in tests not to increase view to high.

Before this commit, some tests (for example dv.nomination.test.ts)
sometimes creates a block with a very high view like 9. It was because
blocks were created so fast and the timestamp of a block is too future
for CodeChain.