Make sync tests not to propagate transactions
These tests use solo consensus. To generate a block with the solo
consensus, there should be a transaction. These tests send a
transaction to a node to make a block. Then the node propagates the
generated block to other nodes. These tests then check the other nodes
whether the received the block or not.

However, if the transaction is propagated before the block, these
tests fail. If the transaction is propagated, other nodes will
generate their own blocks. Then they will reject the block that the
first node generated.